deployment/requirements.txt
torch==2.2.0
transformers==4.39.0
peft==0.10.0
datasets==2.18.0
scikit-learn==1.4.1
pandas==2.2.1
numpy==1.26.4
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pyyaml==6.0.1
matplotlib==3.8.3
seaborn==0.13.2
evaluate==0.4.1
accelerate==0.28.0
nltk==3.8.1
deep_translator==1.11.4
